storing location and activity locally all the time

+++ app/src/pitstop/java/com/pitstop/models/trip/CarActivity.kt
@@ -3,4 +3,4 @@ package com.pitstop.models.trip
 /**
  * Created by Karol Zdebel on 6/4/2018.
  */
-data class CarActivity(val vin: String, val time: Long, val type: Int, val conf: Double)
+data class CarActivity(val vin: String, val time: Long, val type: Int, val conf: Int)
